+++ src/chart/index.js
@@ -1,9 +1,0 @@
-export { default as AreaChart } from "./AreaChart";
-export { default as BarChart } from "./BarChart";
-export { default as LineChart } from "./LineChart";
-export { default as ComposedChart } from "./ComposedChart";
-export { default as PieChart } from "./PieChart";
-export { default as RadarChart } from "./RadarChart";
-export { default as RadialBarChart } from "./RadialBarChart";
-export { default as ScatterChart } from "./ScatterChart";
-