+++ CHANGELOG.md
@@ -10,6 +10,15 @@
 * DataForm 1 field enter reload bug fixed.
 * Added new SelectInput component
 * An error fixed in the name of the prop in the sample code at ReCaptcha.
+* LazyImage added.
+* DateInput bugs fixed
+* FileInput validation fixed.
+* Delay for server requests in search and filtering at DataGrid Component. Search animation added.
+* Countdown component added.
+* Chart components added.
+* [#54](https://github.com/robeio/robe-react-ui/issues/54) fixed.
+* [#50](https://github.com/robeio/robe-react-ui/issues/50) fixed.
+* [#46](https://github.com/robeio/robe-react-ui/issues/46) fixed.
 
 ## 1.1.18
 * upgrade robe-react-commons library.